+++ packages/cli/src/settings.ts
@@ -123,12 +123,14 @@ function _resolveCliSettings(overrides: Partial<CliSettings> = {}): CliSettings 
         `frame,${DEFAULT_REGISTRY_PROVIDER_URL}`,
       registryChainId: process.env.CANNON_REGISTRY_CHAIN_ID || fileSettings.registryChainId || '1',
       registryAddress: process.env.CANNON_REGISTRY_ADDRESS || fileSettings.registryAddress || DEFAULT_REGISTRY_ADDRESS,
-      registryPriority: process.env.CANNON_REGISTRY_PRIORITY || fileSettings.registryPriority || 'onchain',
+      registryPriority: (process.env.CANNON_REGISTRY_PRIORITY ||
+        fileSettings.registryPriority ||
+        'onchain') as CliSettings['registryPriority'],
       etherscanApiUrl: process.env.CANNON_ETHERSCAN_API_URL || fileSettings.etherscanApiUrl || '',
       etherscanApiKey: process.env.CANNON_ETHERSCAN_API_KEY || fileSettings.etherscanApiKey || '',
       quiet: process.env.CANNON_QUIET === 'true' || fileSettings.quiet || false,
     },
-    overrides
+    _.pickBy(overrides)
   );
 
   // Filters out API keys from URLs