+++ packages/website/src/app/learn/guides/debug/page.tsx
@@ -2,7 +2,7 @@ import { DebugPage } from '@/features/Debug/DebugPage';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  title: 'Cannon | Debug',
+  title: 'Cannon | Debugging Tips',
 };
 
 export default function Home() {