+++ packages/website/src/components/MarkdownSpec.tsx
@@ -1,10 +1,12 @@
 'use client';
 
 import { CodePreview } from '@/components/CodePreview';
-import { Text, Code, Heading, Table, Tbody } from '@chakra-ui/react';
+import { Text, Code, Heading, Table } from '@chakra-ui/react';
 import { FC } from 'react';
 import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
 import { headingToId } from '@/helpers/markdown';
+import style from './markdownSpec.module.scss';
 
 interface IDocsPageProps {
   md: string;
@@ -13,6 +15,8 @@ interface IDocsPageProps {
 export const MarkdownSpec: FC<IDocsPageProps> = ({ md }) => {
   return (
     <ReactMarkdown
+      className={style.markdownComp}
+      remarkPlugins={[remarkGfm]}
       components={{
         p: ({ ...props }) => <Text {...props} mt={4} />,
         h1: () => null,
@@ -30,14 +34,16 @@ export const MarkdownSpec: FC<IDocsPageProps> = ({ md }) => {
         h4: () => null,
         a: () => null,
         table: ({ children }) => (
-          <Table m={10} variant="unstyled">
-            <Tbody>{children}</Tbody>
+          <Table m={5} variant="simple">
+            {children}
           </Table>
         ),
         code: ({ inline, className, ...props }) => {
           const lang = className?.replace('language-', '') || 'javascript';
           return inline ? (
-            <Code colorScheme="blackAlpha" variant="solid" {...props} />
+            <Code colorScheme="blackAlpha" variant="solid">
+              {props.children}
+            </Code>
           ) : (
             <CodePreview
               code={(props.children[0] as string) || ''}