+++ packages/hardhat-cannon/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## [2.7.1](https://github.com/usecannon/cannon/compare/v2.7.0...v2.7.1) (2023-09-17)
+
+
+### Bug Fixes
+
+* misc debug mode fixes ([#424](https://github.com/usecannon/cannon/issues/424)) ([2f99df5](https://github.com/usecannon/cannon/commit/2f99df57d599653f86594ac889a0c30f8fb74c21))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @usecannon/builder bumped from ^2.7.0 to ^2.7.1
+    * @usecannon/cli bumped from ^2.7.0 to ^2.7.1
+
+## [2.7.0](https://github.com/usecannon/cannon/compare/v2.6.0...v2.7.0) (2023-09-16)
+
+
+### Features
+
+* **cli:** add anvil custom options to run and build commands ([e9cf80b](https://github.com/usecannon/cannon/commit/e9cf80b385f9cd6fa7d6461d201d200d4090aea9))
+
+
+### Bug Fixes
+
+* gitHeads references from lerna ([ae76eae](https://github.com/usecannon/cannon/commit/ae76eaeda3e83ab14a09198449d0e0f096ee7839))
+
+## [2.6.0](https://github.com/usecannon/cannon/compare/v2.5.6...v2.6.0) (2023-09-16)
+
+
+### Features
+
+* **cli:** add anvil custom options to run and build commands ([e9cf80b](https://github.com/usecannon/cannon/commit/e9cf80b385f9cd6fa7d6461d201d200d4090aea9))
+* migrate cannonfile validation to zod ([#320](https://github.com/usecannon/cannon/issues/320)) ([04ebdf6](https://github.com/usecannon/cannon/commit/04ebdf64f38da38f2076f4cb09b9b1e1c88d721a))
+
+
+### Bug Fixes
+
+* gitHeads references from lerna ([ae76eae](https://github.com/usecannon/cannon/commit/ae76eaeda3e83ab14a09198449d0e0f096ee7839))
+* **hardhat-cannon:** Fixing hardhat compatibility with latest version ([#359](https://github.com/usecannon/cannon/issues/359)) ([6855af9](https://github.com/usecannon/cannon/commit/6855af9bca55492f77ce7a4d37d037455943429e))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @usecannon/builder bumped from ^2.6.1 to ^2.7.0
+    * @usecannon/cli bumped from ^2.6.1 to ^2.7.0
+
 ## [2.5.6](https://github.com/usecannon/cannon/compare/v2.5.5...v2.5.6) (2023-09-13)